+++ lib/haml_lint/lint.rb
@@ -1,11 +1,16 @@
 module HamlLint
   class Lint
-    attr_reader :filename, :line, :message
+    attr_reader :filename, :line, :message, :severity
 
-    def initialize(filename, line, message)
+    def initialize(filename, line, message, severity = :warning)
       @filename = filename
       @line     = line
       @message  = message
+      @severity = severity
+    end
+
+    def error?
+      @severity == :error
     end
   end
 end